updated dht11 simulated values, updated graph

+++ TouchGFX/gui/include/gui/model/Model.hpp
@@ -20,7 +20,7 @@ protected:
     int maxTemperature;
     int minTemperature;
     int humidity;
-    int* temperatureGraphData;
+    short int* temperatureGraphData;
 };
 
 #endif // MODEL_HPP
